+++ scoring_service/config.py
@@ -103,7 +103,7 @@ class Settings(BaseSettings):
     # -------------------------------------------------------------------------
     # Scoring
     # -------------------------------------------------------------------------
-    scoring_cadence_hours: int = Field(
+    scoring_cadence_hours: float = Field(
         default=168,
         description="Hours between scoring rounds (168 = weekly)",
     )
@@ -167,7 +167,7 @@ class Settings(BaseSettings):
         default=500,
         description="Days until a generated VL expires (safety net if service stops publishing)",
     )
-    vl_effective_lookahead_hours: int = Field(
+    vl_effective_lookahead_hours: float = Field(
         default=1,
         description="Hours between VL signing time and the activation timestamp. The signed blob carries this in its 'effective' field so every validator caches the pending blob and activates it simultaneously on the same consensus tick. 0 = activate immediately on fetch.",
     )
